Use both a min-heap and a max-heap for active apps
Strangely I forgot that a heap is not internally ordered.

The code that is used in maintaining a set of active applications only
needs extremes. The minimum time is used to remove stale entries. The
maximum time is used to collect a list of active applications.

This change introduces the use of both a min-heap and a max-heap for
both types of access.

Change-Id: I2bde3e9b895e89035e2e1fdfaa62db0f906a4b88