Remove old copy of server.go and use the library
The fork of the http server used in the router was missing a fair number
of upstream updates and didn't seem to add additional logic or
capability to the router.  Given the recent issue with 204's it seems
like a good idea to revert to the golang library's net/http
implementation.