Omit configuration from /varz output.
Recent exchanges on vcap-dev have resulted in proposals to remove
configuation information from /varz outputs of various components.

Provided CF Runtime stories 65042430 and 65300526 result in the removal
of configuration information from the ruby CF components, it seems to
make sense to begin removing it from the go components as well.

In the case of the go router, omitting configuration will prevent the
broadcase of NATS credentials and the loggregator shared secret.