Refactor route service proxy code
This commit moves all route service code out of the
ProxyRoundTripper.RoundTrip function so that we no longer have issues
with the retry logic.

Two new structs have been created, RouteServiceConfig and
RouteServiceArgs. The config holds all of the crypto/flags/timeouts
needed for route services, and the args struct is a per-request variable
that holds the URL/headers associated with the route service bound, if
any, to the request's destined backend.

[#100238250]