Make vcap component logging work in GoRouter
The common/common.go file sets up a package-wide variable (log), that is
then used for all logging in the vcap component. We found the logging in
that package was not working properly, eventually found the cause to be
this package-wide variable.

On startup, Go will look at the imports in main.go, of which the
common package is one of, and initialize all package-wide variables of
the dependent packages. After this, the main function is run, which is
where the setup for the gosteno.Config is initialized. This means that
the log variable had an incorrect configuration.