During Drain() stop responding to heartbeats
Wait a configurable amount of time before stopping listening.
Abruptly stopping listening may cause latency or lost requests for requests sent to this gorouter from the moment it stops listening.
The drain_wait configuration option can be used to give time to downstream LBs to detect that gorouter is going down soon and steer new traffic away from it before this happens.
If unspecified drain_wait is 0 to avoid changing the current behavior. The cf-release enabling bits will come later.

During Drain() stop responding to heartbeats and wait a configurable amount of time before stopping listening
This is to give time to a downstream load balancer for detecting that gorouter is going down soon and steer new traffic away from it before this happens.