Allow forcing X-Forwarded-Proto Header
There are situations where it's not possible to determine the
appropriate value for this header from the incoming connection (for
example when running behind an Amazon ELB in TCP mode where SSL
termination is handled by the ELB). This therefore adds a config option
to allow forcing the X-Forwarded-Proto header to be https.