Distinguish routekey from uri.
The generation of the routekey treats everything of the URI as
case-insensitive, even the parts which are not (like the path, RFC
3986, Section 6.2.2.1).

Overwriting the uri and then creating the context path from that
causes the system to store a bad context in the registry and later the
VCAP_ID cookie. That breaks session stickiness for applications which
use upper-case characters in their path part.

The separation of routekey from uri prevents storage of an uri with
path converted to lower-case while ensuring that the key does not
violate any assumptions the registry may have regarding it.

Extended the testsuite, demonstrate the new behaviour regarding the
context path of a route.