Refactors handlers to not use Context excessively
This commit refactors most of logic that grabbed metadata directly off
the context into code that grabs metadata off a struct stored in the
context. This should reduce the amount of copies we have to make to
save request metadata and also reduce the amount of type assertions
we have to make.

[#141524603 #144396601]

Signed-off-by: Shash Reddy <sreddy@pivotal.io>