Allow inject HSTS header to response
We want to be able to inject the HSTS header, Strict-Transport-Security
into the response of the gorouter[1]

A new option strict_transport_security_header is added and can be
set to any free string to be inserted in this header, the
operator must set the right format.

The header would be injected only if configured and if the backend
does not return that header.

The implementation extends ProxyResponseWriter to allow inject
any the header, and we add a new custom handler. This would make
it easier to add similar features later.

We add this handler before any other handler that returns a
response, so that the HSTS header is always set (e.g. for 404
unknown route)

[1] https://en.wikipedia.org/wiki/HTTP_Strict_Transport_Security