Allow inject custom headers into response
We want to be able to inject the heades in any response from the
go router. The header would be injected only if the backend
does not return those headers.

One example use case can be add a HSTS header,
Strict-Transport-Security into any response from the gorouter[1]

We allow some new configuration as in the router as:

    http_rewrite:
      inject_response_headers:
      - name: X-A-Header
      value: the-value
      - name: X-A-Header
      value: other-value
      - name: X-Other-Header
      value: some value

This commit adds the basis to implement any rewrite logic on
the headers. We extend ProxyResponseWriter to allow register
new HeaderRewriters.

We add for the time being a InjectHeaderRewriter, that would only
add headers if they don't already exist. Similar HeaderRewriters
can be implemented for delete, append, etc.

A new Handler HTTPRewrite is inserted if there is any `http_rewrite`
configuration. This handler would initialise once HeaderRewriter
during the setup of the server, and attach it to each
ProxyResponseWriter in each request.

[1] https://en.wikipedia.org/wiki/HTTP_Strict_Transport_Security