Add http_rewrite.responses.remove_headers
CF operators have said that it would improve security if they could
specify a list of HTTP headers that should be stripped by gorouter from
responses returned to downstream clients.

So, as operators we want to be able to define a list of headers to
be removed from the response[1], following this configuration
syntax:

    router.http_rewrite.responses.remove_headers:
      description: "Array. List of HTTP headers that will be stripped from responses."

Usage example

    router:
      http_rewrite:
        responses:
          remove_headers:
          - name: X-Vcap-Request-Id

We implement the feature based on the previous code that support
http_rewrites.

We ensure that:
 - `remove_headers` is execute before
   `add_header_if_not_present`, so both can be combined and result
   in a `set_header` operation.
 - `remove_headers` can remove headers added by gorouter like
   `X-Vcap-Request-Id`

[1] https://github.com/cloudfoundry/routing-release/issues/129

fixup! Add http_rewrite.responses.remove_headers