Ensure that header rewriters are canonicalize
The http.Header{} provides methods to ensure that the header is
accessed in a case insentive way[1]

We want to keep this rewrite logic fast so we must do the
canonicalization during setup, not when processing the request.
Because that,  the helpers in `proxy/utils`,
`AddHeaderIfNotPresentRewriter` and `RemoveHeaderRewriter` assume
that you pass a `http.Header` with canonicalized headers. We also
add some tests for to check it works  as expected, but they are
pointless as we rely on `http.Header`.

Instead, the `handler.httpRewriteHandler` must build the right
`http.Header` objects using the methods.

We add a tests to check this.

[1] https://golang.org/pkg/net/http/#Header.Get