Disable aggressive caching for unknown_route resps
If all of the apps serving a particular app fail their healthcheck,
gorouter will respond with a 404.

If there are caching proxies in front of gorouter these may decide to
cache this 404 beyond the point when applications become healthy again.

This can result in the undesirable situation that an application that
has a few seconds of downtime can end up with a cached 404 for hours
(depending on the cache configuration).

By setting `Cache-Control: public,max-age=2` on unknown_route responses
we can instruct any caches in front of gorouter to cache these responses
for a very limited amount of time, to ensure that once apps pass their
healthchecks, 404s do not persist in any caches for very long.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>
Co-authored-by: Richard Towers <richard.towers@digital.cabinet-office.gov.uk>
Co-authored-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>