Lookup handler should not expose LB IP
gorouter requires the Host header to know to which backend to proxy to.

The Host header is optional in HTTP/1.0, so we must explicitly check
that the Host is set.

Also, when some load balancers which are used to load balance gorouters
(e.g. AWS Classic ELB) receive HTTP/1.0 requests without a host headers,
they optimistically add their IP address to the host header.

Therefore to not expose the internal IP address of the load balancer, we
must check that the Host is not also the source of the request.

It is not vali for the Host header to be an IP address, because Gorouter
should not have an IP address as a route.

See https://aws.amazon.com/premiumsupport/knowledge-center/elb-fix-failing-health-checks-alb/
for more information about the Host header.

Closes #256

Signed-off-by: toby lorne <toby@toby.codes>