Set proxy net.Dial KeepAlive to 1 second
When gorouter is configured to keep connections alive
(DisableKeepAlives=false) then it will keep idle TCP connections open to
backends for up to 90 seconds. If the connection is not used after 90
seconds then it is closed; if a new request comes in, and there is a
idle connection to the backend ready, then it will be used instead of a
new TCP connection being made.

The symptoms of this are very similar to this issue:
https://github.com/cloudfoundry/gorouter/issues/209 ; but the cause is
different.

If a backend endpoint closes a connection which has been kept alive at
the same time as gorouter tries to use the connection, then Gorouter
returns a 502 and logs 'backend-endpoint-failed' due to an io.EOF on the
socket.

The underlying TCP connection is managed by net.Dialer which by default
probes the status of the underlying connection every 15s, which
communicates to the recipient that the connection should still be kept
open.

Some apps running in our foundry have keepalives configured to be less
than the 15s default probe interval. Enabling connection keep alives in
our foundry caused gorouter to return 502 (backend-endpoint-failed) for
approximately ~0.01% of requests to backends which had short keepalive
durations. It did not appear to have a substantial effect on apps with
longer keepalive durations, although I have not conclusively tested
this.

This only affects requests which cannot be retried (e.g. POST)

Now that routing-release (since 0.192) defaults keep-alives to backend
connections to be on, it may be sensible to have a shorter probe
interval, to avoid these hard to diagnose networking issues.

This commit configures the default probe interval of the underlying
net.Dialer in the same style as EndpointDialTimeout; it adds a new
configuration parameter EndpointKeepAliveProbeInterval which configures
the underlying net.Dialer KeepAlive parameter, and overrides the default
of 15s to be 1s.

I chose 1s somewhat arbitrarily, but it has the property of being lower
than the 2s default keepalive duration that some of the pathological
apps deployed to our CF have.

This should not adversely affect the performance of gorouter, however that it
will send 1 probe per connection idle connection every one second,
instead of 1 probe per connection every 15 seconds.

Signed-off-by: toby lorne <toby@toby.codes>