Do not use zipkin headers to log directly
The access log takes a list of headers to log, which will include w3c
tracing headers.

In order to ensure these will be able to be logged appropriately,
we pull out the headers to be logged into its own variable

In order to do this, we should simplify the Zipkin handler, and make
a util to deduplicate the headers appropriately.

The Zipkin handler can now be dumb, and not log anything if it is not
enabled, and otherwise present some headers to be logged. This means we
can remove some tests. The only logic we actually care about is that the
headers are uniquely logged, in order.

Signed-off-by: toby lorne <toby@toby.codes>