Increase TLSHandshakeTimeout in proxy test suite
At 500 milliseconds, these tests were often failing:

Proxy Backend Connection Handling when a TLS handshake occurs when the server cert does not match the client when emptyPoolResponseCode503 is false [It] prunes the route
/routing-release/src/code.cloudfoundry.org/gorouter/proxy/proxy_test.go:1020

Proxy Backend Connection Handling when a TLS handshake occurs when the server cert does not match the client when emptyPoolResponseCode503 is true when MaxConns is > 0 [It] prunes the route
/routing-release/src/code.cloudfoundry.org/gorouter/proxy/proxy_test.go:1001

Proxy Backend Connection Handling when a TLS handshake occurs when the server cert does not match the client when emptyPoolResponseCode503 is true [It] prunes the route
/routing-release/src/code.cloudfoundry.org/gorouter/proxy/proxy_test.go:984

with:

```
  Expected
      <int>: 502
  to equal
      <int>: 503
```

Putting it greater than the EndpointTimeout (currently 1s) causes other
tests to fail. May be worth revisiting this test setup at somepoint.