Remove SkipSSLValidation from BackendTLS config
When the gorouter spec property `router.ssl_skip_validation` set to
`true`, it previously caused the TLS-based route integrity feature to
stop working, as mentioned in the description of that property:
"Warning: If this is set to true Gorouter will not validate the backend
identity which could result in misrouting.  For more information see
https://docs.cloudfoundry.org/concepts/http-routing.html#consistency"

This fixes that issue.

https://github.com/cloudfoundry/routing-release/issues/180
[#164261690](https://www.pivotaltracker.com/story/show/164261690)