Make emission of per-request metrics configurable
https://github.com/cloudfoundry/gorouter/issues/159

The metrics latency, latency.component and route_lookup_time are
sent for each request and cause significant load on gorouter and
loggregator. In our measurements the CPU load was reduced by 40%
in both gorouter and doppler VMs. This is a huge gain for large
scale.

Note that the latency metrics is still available at the varz
endpoint.