Support talking to NATS over mTLS
Until now Gorouter has been unable to encrypt its connection to NATS.
Mutual TLS has been added for other components that talk to NATS, but
not to Gorouter.

This commit adds support for configuring the NATS connection with
Mutual TLS. This code is based heavily on the implementation in
route-emitter [1].

The config YAML's structure gets some changes: the username and password
is no longer provided separately for every NATS machine. This isn't a
limitation in real-world Cloud Foundry deployments. Port is repeated for
each NATS machine so that the integration tests can still run multiple
NATS on multiple ports.

[1] https://github.com/cloudfoundry/route-emitter/tree/master/diegonats