Add healthchecker process
Healthchecker is polling gorouter /healthz endpoint. If it returns a
non-200 response healthchecker exits with an error.

When gorouter drain sends USR1, the healthchecker will exit, so
that it does not interfere with gorouter drain process.

Gorouter is sleeping for the configured StartResponseDelayInterval
before it responds to the /healthz endpoint to get its routing table. So
healthchecker is also sleeping for that period plus the 5 second buffer
before polling the /healthz endpoint.

[#179069746](https://www.pivotaltracker.com/story/show/179069746)