Prune endpoints for failed route-services
When a route-service fails, there is a chance the endpoint
at index 0 of an endpointpool is stale and has an out of date
route_service_url. Since this is the endpoint used for all of a pool's
route_service_url calls, all calls to that route will fail, even if
valid endpoints have the correct route service URL. When this happens,
no stale routes can be pruned, since none of the endpoints are ever
tried (the route service call always fails first).

On failure, the routeservice handler will now prune the 0-indexed endpoint from
a request's endpointpool, allowing other endpoints to be used for
the route_service_url until one succeeds, or there are is only one
endpoint left. Any valid endpoints will be recreated, as route_emitter
emits them periodically.

[#179364826](https://www.pivotaltracker.com/story/show/179364826)