Add tests to ensure that we support ALPN-enabled HTTP/1.1 clients
In golang 1.17, ALPN handling was altered slightly in a way that would
break HTTP/1.1 clients using HTTP, if the server only supported 'h2'
in its ALPN protocols. Previously, it would fall back to no-negotiated
mechanism. Golang later added a short-circuit specifically for ALPN
clients supporting http/1.1 with servers supporting h2 to fall-back as
if ALPN was not negotiated.

Modifying router.go to include "http/1.1" isn't strictly required
anymore, but is probably a better thing to do anyway, and it allows
us to validate what protocol was negotiated in our tests.