Fixing 404 response code from gorouter when all app instances are down
There is an open issue in cf/routing-release with id cloudfoundry/routing-release#250.
It introduces a Pool timeout and Pool Updated notions used to prevent pruning empty pools before they timeout.
After the timeout is reached, the pool will be deleted and all coming calls on the related app will result in 404 respose code.
There is also a new gorouter parameter in the config (EmptyPoolTimeout) to set the timeout for pools.

Fixe yaml syntaxe typo in defining PerAppPrometheusHttpMetricsReporting in config.go