Fix default config initialization for websocket dial timeout value
The value is either defined explicitly in the configuration, or derived from the `endpoint_dial_timeout`.
Because the config read from file is applied on top of the values in `Config.defaultConfig`, the
`WebsocketDialTimeout` must not be defined there yet, as it would disable the automatic derivation of the value from a configured `endpoint_dial_timeout` and would change behavor.

The gorouter tests used the `defaultConfig` without `Process()`, so the value was never derived and `WebsocketDialTimeout` remained at `0`, failing some tests.

Enabling it specifically for the `Proxy` test that uses the websocket dial timeout.