Ensure access logs are sent to loggregator when 431s occur.
We only send access logs to loggregator when there is an appID on the
request and sourceInstanceID set on the LogSender. This ensures access
logs can be properly correlated to things developers/operators look at.

Because of gorouter's endpoint retry logic, we don't know for certain
which endpoint handled a request until after it has been handled. Since
the 431 logic short-circuits this, we need to choose an endpoint to
attempt to send the request to, and set that in reqInfo prior to
returning the 431.

This uses the same endpoint logic to as proxy/proxy.go to determine
which endpoint to start with, and accounts for requests using
sticky sessions.