feat(round-tripper): retry requests more often
The round tripper didn't make any effort to understand at which stage a
request failed and consequently only retried requests that failed while
establishing the connection or are idempotent by definition.

This commit drops the existing behaviour of using a classifier and
replaces it with the httptrace library. This allows us to confidently
determine whether data was transmitted to the target and decide whether
we can retry the request or not. A second, optional check is introduced
to retry more often compared to the existing classifiers by allowing all
requests that did not make it past headers transmission to be retried.
Idempotent requests are always retried.

TODO: check if we can get rid of the classifiers all together

See: https://pkg.go.dev/net/http/httptrace
Co-Authored-By: Alexander Lais <alexander.lais@sap.com>
Co-Authored-By: Dominik Froehlich <dominik.froehlich@sap.com>