Use UUID format for VCAP_REQUEST_ID
Switching format would be a breaking change, since it is being parsed as
UUID and should conform with RFC 4122. It is parsed in our components as
GRPC UUID type. It can also be parsed as UUID in any application code.

If none of the trace headers is provided we generate UUID and use it as
VCAP_REQUEST_ID, b3 trace ID and w3c traceparent.

If b3 or w3c headers are provided and conform with UUID format we use
them as VCAP_REQUEST_ID. Otherwise, provided headers are being passed
through and new UUID is generated for VCAP_REQUEST_ID.

Signed-off-by: Marc Paquette <mpaquette@vmware.com>