Add trace identifier to gorouter logs when handling request (#340)
* Consolidate B3, W3C and VCAP request id headers

If zipkin is enabled, either provided or newly generated B3 headers are
used for W3C and VCAP trace headers.

If W3C is enabled, either provided W3C headers are used or B3 headers
are used.

VCAP request id header is always set to B3 headers, W3C headers or new
header is generated if both zipkin and w3c are disabled.

Signed-off-by: Amin Jamali <ajamali@vmware.com>

* Use UUID format for VCAP_REQUEST_ID

Switching format would be a breaking change, since it is being parsed as
UUID and should conform with RFC 4122. It is parsed in our components as
GRPC UUID type. It can also be parsed as UUID in any application code.

If none of the trace headers is provided we generate UUID and use it as
VCAP_REQUEST_ID, b3 trace ID and w3c traceparent.

If b3 or w3c headers are provided and conform with UUID format we use
them as VCAP_REQUEST_ID. Otherwise, provided headers are being passed
through and new UUID is generated for VCAP_REQUEST_ID.

Signed-off-by: Marc Paquette <mpaquette@vmware.com>

* Use logger with trace info in handlers

Switch to logger.Panic from Fatal in handlers. Panic is catched by
PanicCheck and not Fatal.

Signed-off-by: Maria Shaldybin <mariash@vmware.com>

* Update data keys to be consistent across logs

Signed-off-by: Renee Chu <reneec@vmware.com>

---------

Signed-off-by: Amin Jamali <ajamali@vmware.com>
Signed-off-by: Marc Paquette <mpaquette@vmware.com>
Signed-off-by: Maria Shaldybin <mariash@vmware.com>
Signed-off-by: Renee Chu <reneec@vmware.com>
Co-authored-by: Marc Paquette <mpaquette@vmware.com>