put final status code in access logs, not 1XX
A 1xx Informational status code means that the server has received the
request and is continuing the process. A 1xx status code is purely
temporary and is given while the request processing continues.

After a response with a 1XX status code, there should always be a
reponse after it. We shouldn't hide the status of the "real" response.

This change makes gorouter backwards compatible with how the access logs
+ 1XX responses worked before we bumped to go 1.20.