Continue to forward relevant headers
by removing hop-by-hop headers that could interfere with gorouter.

Context:
    - https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Connection
    - https://nathandavison.com/blog/abusing-http-hop-by-hop-request-headers

The following list of keywords are removed from Connection header, so
that gorouter can continue to utilize those headers when they are set
and will continue to forward them down to the relavant handlers.

nits: Use io.ReadAll instead of ioutil since that's been the default as
of Go1.16