Add localhost-only listener for /routes endpoint
This is part of a larger effort to remove unencrypted
non-localhost TCP listeners from routing-release. Currently
the listener for /routes on `router.status.port` listens on
all interfaces, and sits alongside the deprecated /varz + /healthz
endpoints, as well as the LB healthcheck endpoint.

To work toward the desired state, we're adding a localhost-only
listener for /routes, and will turn `router.status.port` to only a
LB-healthcheck endpoint.

[#186497108](https://www.pivotaltracker.com/story/show/186497108)