Optimistic route registry locks (#378)
* perf(route-registry): benchmark for concurrent route register + lookup

The benchmark checks the concurrent access and locking behavior for
the route registry. This benchmark can be used to assess improvements
in the locking characteristics and concurrency of the route registry.

* perf(route-registry): optimistic read on route register

When processing a register message, lock the route registry
for read-only access first to check, whether an update is necessary.
Only if it is necessary, get a write lock and write the update, as
before.

This ensures that for the usual case of an app's route registration
just being repeated by route-emitter, this does not force a write
lock and will not interfere with resolving routes for other requests.