Add more descriptive comments in roundrobin
- Calculate `curIsLocal` before incrementing `curIdx` for legibility
- Add more comments

I tried to make the code more legible by adding a `nextIndex` variable
and not pre-incrementing the index before the end of the for loop but it
added a bit more of a performance hit than I would have liked. Instead I
settled on adding a comment to hopefully make the logic clearer.