WIP NEEDING FIXES: revert the custom ReverseProxy.Rewrite logic
In these PRs, we added custom logic for handling 100-Continue
requests to avoid a security issue that has since been resolved
upstream:

https://github.com/cloudfoundry/gorouter/pull/418
https://github.com/cloudfoundry/gorouter/pull/419
https://github.com/cloudfoundry/gorouter/pull/420
https://github.com/cloudfoundry/gorouter/pull/424

This commit reverts those changes to the proxy logic prior to #418,
while keeping relevant tests added since, to ensure the issue is not
occurring.

FIXME: However, there is currently a race condition in golang's
http.Server that occurs when calling Flush() on requests when executing
them in the fashion we do in the "does not cause DoS of subsequent
requests" test. This is likely fixed in Golang 1.23, so pausing this
work until Gorouter is using Go 1.23.x.