Syslog Improvements (2) (#459)
* feat: make syslog logging useable

* do not log an empty new line when writing to syslog
* allow operator to configure truncation on syslog

* fix: allow syslog-only logging

* feat: replace log/syslog with internal version

The go standard library implementation of log/syslog has been abandoned
and is no longer maintained. In our tests with syslog logging in
gorouter we noticed a significant performance penalty over file based
logging. However, with file based logging we frequently observe broken
logs because logrotate interferes with log reading / writing.

This commit inlines the relevant file from log/syslog and refactors the
implementation to be on-par with the file based logging.

Co-authored-by: Alexander Lais <alexander.lais@sap.com>

* fix: rename testState.newRequest -> testState.newGetRequest

* test: add integration test for syslog access logs

* test: check errors

---------

Co-authored-by: Alexander Lais <alexander.lais@sap.com>